rabbit/rpc/requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pika==1.2.0
python-dotenv==0.20.0
Werkzeug==2.0.1
